requirements.txt
requests==2.26.0
biopython==1.79

